confirmation for startLayout

+++ custom-scripts/customStartLayout.ps1
@@ -1,10 +1,17 @@
-$destination = $home + "\AppData\Local\Microsoft\Windows\Shell"#
-$source = $PSScriptRoot + "\LayoutModification.xml"
 
-Copy-Item $source -Destination $destination
+$confirmation = Read-Host "This will change your Taskbar and Start-Menu Layout. For changes edit the layoutModification.xml in this directory $PSScriptRoot (Do you agree?) [y/N]"
+if ($confirmation.Length -ne 0) {
+  if ($confirmation.Substring(0, 1).ToLower() -eq 'y') {
+    $destination = $home + "\AppData\Local\Microsoft\Windows\Shell"#
+    $source = $PSScriptRoot + "\LayoutModification.xml"
 
-$path = "HKCU:\SOFTWARE\Microsoft\Windows\CurrentVersion\CloudStore\Store\Cache"
+    Copy-Item $source -Destination $destination
 
-Remove-Item $path -Force -Recurse
+    $path = "HKCU:\SOFTWARE\Microsoft\Windows\CurrentVersion\CloudStore\Store\Cache"
 
-Stop-Process -ProcessName explorer
+    Remove-Item $path -Force -Recurse
+
+    Stop-Process -ProcessName explorer
+  }
+}
+
